refactor(tg-py2ts): cleanups
